Generate a new "visitor" user
What? Why?
==========

Instead of asking someone to immediately create an account -- because,
let's be honest, who wants to create an account right off the bat for
a giant meme database? -- just create a fake/ghost/visitor user that
can eventually be evolved into a "normal" user.

Down the road this can send an email to sign someone in via a secret link.

How?
====

Create the temp user, with a temp name, let people use the thing.

More
====

Using rspec and a PORO _feature object_ we can write our tests
in a manner that reads a little easier.