Image uploaded via Shubox & passed to Images#new
Why?
====

Using turbo frames, react to a successfully uploaded file via Shubox.js
and open the ImagesController#new modal to then allow for the creation of
the new image.

What?
=====

Progressively enhance this. `ImagesController#new` should work on its own as
long as it has a proper image passed to it with the querystring.

Add proper I18n translations to en.yml

Allow for both uploading a file, and providing a url to an image.